Move simple styled search bar to a separate component
